feat: Added OIDC authentication for users
Moved the Authenticator to a separate file

Added an eventlistener for when the access token is about to expire

After logging in the url is cleaned up.

Updated the README.md
