+++ frontend/README.md
@@ -1,50 +1,28 @@
-# React + TypeScript + Vite
+# NL Portal Configuration Panel - Frontend
 
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
+This frontend provides a user interface to configure the NL Portal application.
 
-Currently, two official plugins are available:
+The resulting configuration is stored in a database where the NL Portal can retrieve it.
 
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
+## Development
 
-## Expanding the ESLint configuration
+To contribute to this repository first [clone](https://git-scm.com/docs/git-clone) to your device.
+Make sure to [install pnpm](https://pnpmpkg.com/getting-started/install).
 
-If you are developing a production application, we recommend updating the configuration to enable type aware lint rules:
+This is a React application runninng on Vite
 
-- Configure the top-level `parserOptions` property like this:
+### Installing dependencies:
 
-```js
-export default tseslint.config({
-  languageOptions: {
-    // other options...
-    parserOptions: {
-      project: ['./tsconfig.node.json', './tsconfig.app.json'],
-      tsconfigRootDir: import.meta.dirname,
-    },
-  },
-})
-```
+Install dependencies by running the command `pnpm install` from the frontend folder.
 
-- Replace `tseslint.configs.recommended` to `tseslint.configs.recommendedTypeChecked` or `tseslint.configs.strictTypeChecked`
-- Optionally add `...tseslint.configs.stylisticTypeChecked`
-- Install [eslint-plugin-react](https://github.com/jsx-eslint/eslint-plugin-react) and update the config:
+### Configure the application
 
-```js
-// eslint.config.js
-import react from 'eslint-plugin-react'
+The default configuration added is based on usage of the project [nl-portal-docker-compose](https://github.com/nl-portal/nl-portal-docker-compose).
 
-export default tseslint.config({
-  // Set the react version
-  settings: { react: { version: '18.3' } },
-  plugins: {
-    // Add the react plugin
-    react,
-  },
-  rules: {
-    // other rules...
-    // Enable its recommended rules
-    ...react.configs.recommended.rules,
-    ...react.configs['jsx-runtime'].rules,
-  },
-})
-```
+When you want to use a different configuration you can adjust the [config.js](public/config.js)
+or set up the environment variables used on the [config-template](public/config.template.js).
+
+### Starting the application
+
+After installing the dependencies and making sure the configuration is correct, start the project with `pnpm dev` from the frontend folder.
+
